remove the RemoveQuote API, and the remove quote p2p messages
We think that the RemoveQuote API is not good to offer to clients,
because ultimately, once they have signed a quote, nothing can
guarantee that someone else won't match it, and in principle,
someone can fork this code, and not remove quotes until they are
actually expired. Then they could still be matched, even if we
returned a "success" status on a RemoveQuote request.

In order to get users used to the idea that they have to cancel
their quotes by sending a mobilecoin transaction, rather than just
calling this API, we think we just shouldn't have this API.

Because we don't have this API, we also don't need to have this in
the P2P gossip channels either -- all the P2P nodes will see that
the quote expired because they are syncing the ledger, and that's
their source of truth for whether it has truly expired.

We still send SciQuoteRemoved messages over the LiveUpdates channel.
(Maybe we should call it SciQuoteExpired?)

This added some more tests that had to change after this change,
and adjusted an e2e test that was testing that RemoveQuote was working.

We also added a new status code for `QUOTE_IS_STALE` to the SubmitQuote
API, because otherwise stale quotes got back the non-specific `OTHER`
code.